requirements.txt
aiohttp==3.9.0
aiosignal==1.3.1
anyio==4.0.0
asgiref==3.7.2
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==23.1.0
Authlib==1.2.1
bcrypt==4.0.1
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloud-sql-python-connector==1.4.3
colorama==0.4.6
connexion==3.0.1
contextlib2==21.6.0
cryptography==41.0.5
exceptiongroup==1.2.0
extras==1.0.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
flask-marshmallow==0.15.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.0
google-auth==2.23.4
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
importlib-metadata==6.8.0
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
packaging==23.2
pg8000==1.30.3
protobuf==4.21.12
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.12.0
rsa==4.9
schema==0.7.5
scramp==1.4.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.32.0.post1
typing_extensions==4.8.0
urllib3==2.0.7
Werkzeug==3.0.1
yarl==1.9.2
zipp==3.17.0
